// src/types/mopidy.d.ts
export declare class Mopidy {
  // ----------------- MOPIDY.JS-SPECIFIC API -----------------

  /**
   * Mopidy.js is a JavaScript library for controlling a Mopidy music server.
   *
   * The library makes Mopidy's core API available from browsers and Node.js
   * programs, using JSON-RPC over a WebSocket to communicate with Mopidy.
   *
   * This library is the foundation of most Mopidy web clients.
   */
  constructor(options: Mopidy.Options);
  /**
   * Explicit connect function for when autoConnect:false is passed to
   * constructor.
   */
  connect(): Promise<void>;
  /**
   * Close the WebSocket without reconnecting. Letting the object be garbage
   * collected will have the same effect, so this isn't strictly necessary.
   */
  close(): Promise<void>;

  // ----------------- EVENT SUBSCRIPTION -----------------
  
  on<K extends keyof Mopidy.StrictEvents>(
    name: K, listener: Mopidy.StrictEvents[K]
  ): this;

  off(): void;
  off<K extends keyof Mopidy.StrictEvents>(
    name: K, listener: Mopidy.StrictEvents[K]
 ): this;

  // ----------------- CORE API -----------------

  /**
   * Manages everything related to the list of tracks we will play. See
   * TracklistController. Undefined before Mopidy connects.
   */
  tracklist: Mopidy.Core.TracklistController | undefined;
  /**
   * Manages playback state and the current playing track. See
   * PlaybackController. Undefined before Mopidy connects.
   */
  playback: Mopidy.Core.PlaybackController | undefined;
  /**
   * Manages the music library, e.g. searching and browsing for music. See
   * LibraryController. Undefined before Mopidy connects.
   */
  library: Mopidy.Core.LibraryController | undefined;
  /**
   * Manages stored playlists. See PlaylistsController. Undefined before
   * Mopidy connects.
   */
  playlists: Mopidy.Core.PlaylistsController | undefined;
  /**
   * Manages volume and muting. See MixerController. Undefined before Mopidy
   * connects.
   */
  mixer: Mopidy.Core.MixerController | undefined;
  /**
   * Keeps record of what tracks have been played. See HistoryController.
   * Undefined before Mopidy connects.
   */
  history: Mopidy.Core.HistoryController | undefined;

  /**
   * Get list of URI schemes we can handle
   */
  getUriSchemes(): Promise<string[]>;
  /**
   * Get version of the Mopidy core API
   */
  getVersion(): Promise<string>;
}

export declare namespace Mopidy {
  type ValueOf<T> = T[keyof T];

  interface Options {
    /**
     * URL used when creating new WebSocket objects.
     *
     * In a browser environment, it defaults to
     * ws://${document.location.host}/mopidy/ws. If the current page is served
     * over HTTPS, it defaults to using wss:// instead of ws://.
     *
     * In a non-browser environment, where document.location isn't available, it
     * defaults to ws://localhost/mopidy/ws.
     */
    webSocketUrl: string;
    /**
     * Whether or not to connect to the WebSocket on instance creation. Defaults
     * to true.
     */
    autoConnect?: boolean;
    /**
     * The minimum number of milliseconds to wait after a connection error before
     * we try to reconnect. For every failed attempt, the backoff delay is doubled
     * until it reaches backoffDelayMax. Defaults to 1000.
     */
    backoffDelayMin?: number;
    /**
     * The maximum number of milliseconds to wait after a connection error before
     * we try to reconnect. Defaults to 64000.
     */
    backoffDelayMax?: number;
    /**
     * If set, this object will be used to log errors from Mopidy.js. This is
     * mostly useful for testing Mopidy.js. Defaults to console.
     */
    console?: Console;
    /**
     * An existing WebSocket object to be used instead of creating a new
     * WebSocket. Defaults to undefined.
     */
    webSocket?: WebSocket;
  }
  type URI = string;
  type PlaybackState = "playing" | "paused" | "stopped";
  type QueryField =
    | "uri"
    | "track_name"
    | "album"
    | "artist"
    | "albumartist"
    | "composer"
    | "performer"
    | "track_no"
    | "genre"
    | "date"
    | "comment"
    | "any";
  type Query = { [key in QueryField]?: string[] };
  interface StrictEvents extends Core.CoreListener {
    /**
     * Client state
     *
     * You can get notified about when the Mopidy.js client is connected to the
     * server and ready for method calls, when it's offline, and when it's
     * trying to reconnect to the server by looking at the events
     */
    "state:online": () => void;
    "state:offline": () => void;
    reconnectionPending: ({ timeToAttempt }: { timeToAttempt: number }) => void;
    reconnecting: () => void;
    /**
     * The client state events are also emitted under the aggregate event named
     * state.
     */
    state: (args?: unknown) => void;

    /**
     * WebSocket events
     *
     * You can introspect what happens internally on the WebSocket by looking at
     * the events.
     *
     * Of course, you can also do this using the web developer tools in any
     * modern browser.
     */
    "websocket:close": any;
    "websocket:error": any;
    "websocket:incomingMessage": any;
    "websocket:open": any;
    "websocket:outgoingMessage": any;
  }
// ...
  namespace Core {
// ...
    interface PlaybackController {
      /**
       * Play the given track, or if the given `tl_track` and `tlid` is None,
       * play the currently active track.
       *
       * Note that the track *must* already be in the tracklist.
       */
      play({
        track,
        tlid,
      }: {
        track?: models.TlTrack;
        tlid?: number;
      }): Promise<void>;

      /**
       * Change to the next track.
       *
       * The current playback state will be kept. If it was playing, playing will
       * continue. If it was paused, it will still be paused, etc.
       */
      next(): Promise<void>;

      /**
       * Change to the previous track.
       *
       * The current playback state will be kept. If it was playing, playing will
       * continue. If it was paused, it will still be paused, etc.
       */
      previous(): Promise<void>;

      /**
       * Stop playing.
       */
      stop(): Promise<void>;

      /**
       * Pause playback.
       */
      pause(): Promise<void>;

      /**
       * If paused, resume playing the current track.
       */
      resume(): Promise<void>;

      /**
       * Seeks to time position given in milliseconds.
       */
      seek({
        time_position
      }: {
        /**
         * time position in milliseconds
         */
        time_position: number
      }): Promise<boolean>;

      // ----------------- CURRENT TRACK -----------------

      /**
       * Get the currently playing or selected track.
       */
      getCurrentTlTrack(): Promise<models.TlTrack | null>;

      /**
       * Get the currently playing or selected track.
       *
       * Extracted from `get_current_tl_track()` for convenience.
       */
      getCurrentTrack(): Promise<models.Track | null>;

      /**
       * Get the current stream title or None.
       */
      getStreamTitle(): Promise<string | null>;

      /**
       * Get time position in milliseconds.
       */
      getTimePosition(): Promise<number | null>;

      // ----------------- PLAYBACK STATES -----------------

      /**
       * Get The playback state.
       */
      getState(): Promise<PlaybackState>;

      /**
       * Set the playback state. See:
       *  https://docs.mopidy.com/en/latest/api/core/#mopidy.core.PlaybackController.set_state
       * for possible states and transitions
       */
      setState({ new_state }: { new_state: PlaybackState }): Promise<void>;
    }

    // https://docs.mopidy.com/en/latest/api/core/#library-controller
    interface LibraryController {
      /**
       * Browse directories and tracks at the given uri.
       *
       * uri is a string which represents some directory belonging to a backend.
       * To get the intial root directories for backends pass None as the URI.
       *
       * returns a list of `mopidy.models.Ref` objects for the directories and
       * tracks at the given uri.
       *
       * The Ref objects representing tracks keep the track's original URI. A
       * matching pair of objects can look like this:
       *
       *    Track(uri='dummy:/foo.mp3', name='foo', artists=..., album=...)
       *    Ref.track(uri='dummy:/foo.mp3', name='foo')
       *
       * The Ref objects representing directories have backend specific URIs.
       * These are opaque values, so no one but the backend that created them
       * should try and derive any meaning from them. The only valid exception to
       * this is checking the scheme, as it is used to route browse requests to
       * the correct backend.
       *
       * For example, the dummy library's /bar directory could be returned like
       * this:
       *
       *    `Ref.directory(uri='dummy:directory:/bar', name='bar')`
       */
      browse({
        uri
      }: {
        /**
         * URI to browse
         */
        uri: URI
      }): Promise<models.Ref<any>[]>;

      /**
       * Search the library for tracks where `field` contains `values`.
       *
       * `field` can be one of `uri`, `track_name`, `album`, `artist`, `albumartist`,
       * `composer`, `performer`, `track_no`, `genre`, `date`, `comment`, or `any`.
       *
       * If `uris` is given, the search is limited to results from within the URI
       * roots. For example passing `uris=['file:']` will limit the search to the
       * local backend.
       *
       * Examples:
       *
       *     # Returns results matching 'a' in any backend
       *     search({'any': ['a']})
       *
       *     # Returns results matching artist 'xyz' in any backend
       *     search({'artist': ['xyz']})
       *
       *     # Returns results matching 'a' and 'b' and artist 'xyz' in any
       *     # backend
       *     search({'any': ['a', 'b'], 'artist': ['xyz']})
       *
       *     # Returns results matching 'a' if within the given URI roots
       *     # "file:///media/music" and "spotify:"
       *     search({'any': ['a']}, uris=['file:///media/music', 'spotify:'])
       *
       *     # Returns results matching artist 'xyz' and 'abc' in any backend
       *     search({'artist': ['xyz', 'abc']})
       */
      search({
        query,
        uris,
        exact,
      }: {
        /**
         * one or more queries to search for
         */
        query: Query;
        /**
         * zero or more URI roots to limit the search to
         */
        uris?: string[];
        /**
         * if the search should use exact matching
         */
        exact?: boolean;
      }): Promise<models.SearchResult[]>;

      /**
       * Lookup the given URIs.
       *
       * If the URI expands to multiple tracks, the returned list will contain them all.
       */
      lookup({
        uris
      }:{
        /**
         * A list of URI's
         */
        uris: string[]
      }): Promise<{ [index: string]: models.Track[] }>;

      /**
       *
       * Refresh library. Limit to URI and below if an URI is given.
       */
      refresh({ uri }: { uri?: string }): Promise<void>

      /**
       * Lookup the images for the given URIs
       *
       * Backends can use this to return image URIs for any URI they know about be
       * it tracks, albums, playlists. The lookup result is a dictionary mapping
       * the provided URIs to lists of images.
       *
       * Unknown URIs or URIs the corresponding backend couldn't find anything for
       * will simply return an empty list for that URI.
       */
      getImages({
        uris
      }:{
        /**
         * A list of URI's
         */
        uris: string[]
      }): Promise<{ [index: string]: models.Image[] }>;
    }
// ...
    interface HistoryController {
      /**
       * Get the track history.
       *
       * The timestamps are milliseconds since epoch.
       */
      getHistory(): Promise<{ [index: string]: models.Ref<any>[] }>;

      /**
       * Get the number of tracks in the history.
       */
      getLength(): Promise<number>;
    }
  }
}